src/backend/pyproject.toml
[project]
name = "se-project-steven-zhou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.5.0",
    "dotenv>=0.9.9",
    "fastapi>=0.128.2",
    "ollama>=0.6.1",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlmodel>=0.0.32",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "torchtune>=0.6.1",
    "typing>=3.10.0.0",
    "uvicorn>=0.40.0",
    "weaviate>=0.1.2",
    "weaviate-client>=4.20.4",
]
